refactor: Game 클래스 삭제, CarsReader, RoundReader클래스로 분리
1. CarsReader 클래스 정의
2. RoundReader 클래스 정의
3. utils 하위 패키지 정리(validator, reader)
