+++ src/main/java/racingcar/utils/reader/CarsReader.java
@@ -1,31 +1,18 @@
-package racingcar;
+package racingcar.utils.reader;
 
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
 import camp.nextstep.edu.missionutils.Console;
-import racingcar.utils.Const;
 import racingcar.domain.Car;
-import racingcar.domain.Race;
-import racingcar.utils.NameValidator;
-import racingcar.utils.RoundValidator;
+import racingcar.utils.Const;
+import racingcar.utils.validator.NameValidator;
 
-public class Game {
-	public static void play() {
-		List<Car> cars = readCars();
-		int round = readRound();
-
-		Race race = new Race(cars, round);
-
-		race.start();
-		race.printWinners();
-	}
-
-	private static List<Car> readCars() {
+public class CarsReader {
+	public static List<Car> readCars() {
 		System.out.println(Const.CAR_INPUT_MESSAGE);
 		List<Car> result;
-
 		String userInput = Console.readLine();
 		List<String> names = Arrays.asList(userInput.split(Const.NAME_SEPARATOR));
 
@@ -44,20 +31,6 @@ public class Game {
 
 		for (String name : names) {
 			result.add(new Car(NameValidator.validate(name)));
-		}
-
-		return result;
-	}
-
-	private static int readRound() {
-		System.out.println(Const.ROUND_INPUT_MESSAGE);
-		String userInput = Console.readLine();
-		int result;
-		try {
-			result = RoundValidator.validate(userInput);
-		} catch (IllegalArgumentException e) {
-			System.out.println(e.getMessage());
-			return readRound();
 		}
 
 		return result;
